Añadiendo integración de Gitlab con NGitlab

+++ TFG/Application/Services/GitlabIntegration/Dtos/GitlabAddMembersToProjectDto.cs
@@ -4,7 +4,7 @@ namespace TFG.Application.Services.GitlabIntegration.Dtos
 {
 	public record GitlabAddMembersToProjectDto
 	{
-		public required int Id { get; set; } // The Id of the project
+		public required long Id { get; set; } // The Id of the project
 		public required string UserId { get; set; } // Can receive multiple user ids separated by commas
 		public required GitlabAcessLevel AccessLevel { get; set; } // The access level of the user
 	}
