+++ TFG/Application/Services/GitlabIntegration/Dtos/GitlabCreateProjectResponseDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace TFG.Application.Services.GitlabIntegration.Dtos
 {
-	public record GitlabCreateProjectResponseDto(int Id, string Name, string Description);
+	public record GitlabCreateProjectResponseDto(long Id, string Name, string Description);
 }
